:recycle: Add version id to file name when downloaded
This adds the version kf_id to the file name when downloading so that
the user gets a file that is named specifically for that version.
Otherwise, the user may download a file again either resulting in copies
of the same file which may be difficult to distinguish versions from.